Added new API methods: getEventsByMonth, getEventsByGroupId, getEventsByDay. All three of these are completed.
Modified some existing API calls and provided a way to use the APIs via internal pages in the application by wrapping the core logic in a function and using SCRIPT_FILENAME to determine how the file was being accessed.

Date picker is now displayed on the page permanently so that selecting multiple days is easier and the picker does not keep opening and closing.